+++ src/musikdroid/app/src/main/java/io/casey/musikcube/remote/ui/tracks/constant/Track.kt
@@ -6,6 +6,7 @@ object Track {
         const val SELECTED_ID = "extra_selected_id"
         const val TITLE_ID = "extra_title_id"
         const val CATEGORY_VALUE = "extra_category_value"
+        const val FRAGMENT_ARGUMENTS = "extra_fragment_arguments"
     }
 
     object RequestCode {
